refactor(eve): harness - extract tool model output normalization into its own module
tools.ts is the tool-set assembly module; the content-parts change had
grown a second concern inside it — a ~165-line validation subsystem for
authored toModelOutput results. It moves to tool-model-output.ts next to
the serialization error machinery it throws.

Exporting the normalizer also deletes the test ceremony: the shape
matrix no longer builds a HarnessToolMap and fishes the toModelOutput
closure out of the SDK tool with casts per case — it calls the function.
tools.test.ts keeps one wiring test proving buildToolSet applies the
normalizer.
