+++ packages/eve/src/harness/tools.test.ts
@@ -5,7 +5,7 @@ import { ContextContainer, contextStorage } from "#context/container.js";
 import { SessionKey, type Session } from "#context/keys.js";
 import { SCHEDULE_APP_AUTH } from "#channel/schedule-auth.js";
 import { always, never, once } from "#public/tools/approval/approval-helpers.js";
-import { toolOutput, toolOutputPart } from "#public/tools/output-builders.js";
+
 import type { RuntimeModelReference } from "#runtime/agent/bootstrap.js";
 import {
   WEB_SEARCH_ANTHROPIC_OUTPUT_SCHEMA,
@@ -679,165 +679,6 @@ describe("buildToolSet", () => {
         },
       ],
     });
-  });
-
-  it("accepts outputs built with the toolOutput builders", async () => {
-    const tools: HarnessToolMap = new Map<string, HarnessToolDefinition>(
-      (
-        [
-          ["as_text", () => toolOutput.text("visible")],
-          ["as_json", () => toolOutput.json({ summary: "ok" })],
-        ] as const
-      ).map(([name, toModelOutput]) => [
-        name,
-        {
-          description: "Project output.",
-          execute: async () => ({ ok: true }),
-          inputSchema: jsonSchema({}),
-          name,
-          toModelOutput,
-        },
-      ]),
-    );
-
-    const result = buildToolSet({ tools });
-
-    await expect(
-      projectSdkToolOutput({ output: { ok: true }, tool: result.as_text }),
-    ).resolves.toEqual({ type: "text", value: "visible" });
-    await expect(
-      projectSdkToolOutput({ output: { ok: true }, tool: result.as_json }),
-    ).resolves.toEqual({ type: "json", value: { summary: "ok" } });
-  });
-
-  it("accepts content parts built with the toolOutputPart builders", async () => {
-    const tools: HarnessToolMap = new Map<string, HarnessToolDefinition>([
-      [
-        "screenshot",
-        {
-          description: "Capture a screenshot.",
-          execute: async () => ({ ok: true }),
-          inputSchema: jsonSchema({}),
-          name: "screenshot",
-          toModelOutput: () =>
-            toolOutput.content([
-              toolOutputPart.text("Screenshot:"),
-              toolOutputPart.file("aGVsbG8=", { mediaType: "image/png" }),
-            ]),
-        },
-      ],
-    ]);
-
-    const result = buildToolSet({ tools });
-
-    await expect(
-      projectSdkToolOutput({ output: { ok: true }, tool: result.screenshot }),
-    ).resolves.toEqual({
-      type: "content",
-      value: [
-        { type: "text", text: "Screenshot:" },
-        { type: "file", data: { type: "data", data: "aGVsbG8=" }, mediaType: "image/png" },
-      ],
-    });
-  });
-
-  it.each([
-    [
-      "raw Uint8Array payload",
-      { type: "data", data: new Uint8Array([1, 2, 3]) },
-      /base64 string, received raw bytes/u,
-    ],
-    ["raw untagged bytes", new Uint8Array([1, 2, 3]), /base64 string, received raw bytes/u],
-    [
-      "url file-data tag",
-      { type: "url", url: "https://example.com/a.png" },
-      /"url" is not supported yet/u,
-    ],
-    [
-      "reference file-data tag",
-      { type: "reference", reference: "file_abc" },
-      /"reference" is not supported yet/u,
-    ],
-    ["text file-data tag", { type: "text", text: "inline" }, /"text" is not supported yet/u],
-    ["untagged string data", "aGVsbG8=", /tagged object/u],
-  ] as const)("rejects content file parts with %s", async (_label, data, expected) => {
-    const tools: HarnessToolMap = new Map<string, HarnessToolDefinition>([
-      [
-        "screenshot",
-        {
-          description: "Capture a screenshot.",
-          execute: async () => ({ ok: true }),
-          inputSchema: jsonSchema({}),
-          name: "screenshot",
-          toModelOutput: () => ({
-            type: "content" as const,
-            value: [{ type: "file" as const, data: data as never, mediaType: "image/png" }],
-          }),
-        },
-      ],
-    ]);
-
-    const result = buildToolSet({ tools });
-
-    await expect(
-      projectSdkToolOutput({
-        output: { ok: true },
-        tool: result.screenshot,
-        toolCallId: "call_screenshot",
-      }),
-    ).rejects.toThrow(expected);
-  });
-
-  it("rejects empty content part arrays with the toModelOutput error identity", async () => {
-    const tools: HarnessToolMap = new Map<string, HarnessToolDefinition>([
-      [
-        "screenshot",
-        {
-          description: "Capture a screenshot.",
-          execute: async () => ({ ok: true }),
-          inputSchema: jsonSchema({}),
-          name: "screenshot",
-          toModelOutput: () => ({ type: "content" as const, value: [] }),
-        },
-      ],
-    ]);
-
-    const result = buildToolSet({ tools });
-
-    await expect(
-      projectSdkToolOutput({
-        output: { ok: true },
-        tool: result.screenshot,
-        toolCallId: "call_screenshot",
-      }),
-    ).rejects.toThrow(
-      'Tool "screenshot" call "call_screenshot" returned a non-JSON-serializable model output. ' +
-        'Expected content model output "value" to be a non-empty array.',
-    );
-  });
-
-  it("rejects unknown content part types", async () => {
-    const tools: HarnessToolMap = new Map<string, HarnessToolDefinition>([
-      [
-        "screenshot",
-        {
-          description: "Capture a screenshot.",
-          execute: async () => ({ ok: true }),
-          inputSchema: jsonSchema({}),
-          name: "screenshot",
-          toModelOutput: () => ({
-            type: "content" as const,
-            value: [{ type: "media", data: "aGVsbG8=", mediaType: "image/png" } as never],
-          }),
-        },
-      ],
-    ]);
-
-    const result = buildToolSet({ tools });
-
-    await expect(
-      projectSdkToolOutput({ output: { ok: true }, tool: result.screenshot }),
-    ).rejects.toThrow('Expected content part type to be "text" or "file".');
   });
 
   it("passes valid text toModelOutput values through", async () => {
